User validations

+++ src/controllers/UserController.js
@@ -1,7 +1,8 @@
 const { send, json } = require('micro')
-const { hash } = require('bcrypt')
+const { hash, compare } = require('bcrypt')
 
 const { User } = require('../models/User')
+const SchemaUser = require('../SchemaUser')
 
 module.exports = {
     async index(req, res) {
@@ -28,10 +29,31 @@ module.exports = {
 
     async create(req, res) {
         try {
-            let { name, email, password, address } = await json(req)
+            let { name, email, password, repeat_password, address } = await json(req)
+
+            const users = await User.query()
+            const foundUser = await users.find(user => {
+                if (user.email == email) return user
+            })
+
+            if (foundUser) return send(res, 400, { error: 'The email is being used by another user' })
+
+            const schema = SchemaUser.userSchema
+            const { error } = schema.validate({
+                name: name,
+                email: email,
+                password: password,
+                repeat_password: repeat_password,
+                address: address
+            })
+
+            if (error) {
+                const message = error.message
+
+                return send(res, 400, { message })
+            }
 
             password = await hash(password, 8)
-
             await User.query().insert({
                 name,
                 email,
@@ -39,11 +61,10 @@ module.exports = {
                 address
             })
 
-            return send(res, 201, { success: 'Usuário cadastrado com sucesso!' })
+            return send(res, 201, { success: 'User successfully registered' })
 
         } catch (error) {
-            console.log(error)
-            send(res, 500, { error: 'Erro no servidor.' })
+            send(res, 500, { error: 'Server error' })
         }
 
     },
@@ -54,7 +75,7 @@ module.exports = {
 
             let user = await User.query().findById(id)
 
-            if (!user) return send(res, 204, { error: 'Usuário não encontrado' })
+            if (!user) return send(res, 204, { error: 'User not found' })
 
             user = {
                 name: user.name,
@@ -67,20 +88,41 @@ module.exports = {
             return send(res, 200, { user })
 
         } catch (error) {
-            send(res, 500, { error: 'Erro no servidor.' })
+            send(res, 500, { error: 'Server error' })
         }
     },
 
     async update(req, res) {
         try {
             const { id } = req.params
-            let { name, email, address } = await json(req)
+            let { name, email, address, password } = await json(req)
 
             let user = await User.query().findById(id)
+            if (!user) return send(res, 204, { error: 'User not found' })
 
-            if (!user) return send(res, 204, { error: 'Usuário não encontrado' })
+            const schema = SchemaUser.userSchema
+            const { error } = schema.validate({
+                name: name,
+                email: email,
+                password: password,
+                address: address
+            })
 
-            if(foundUser) return send(res, 400, {error: 'O email está em uso por outro usuário '})
+            if (error) {
+                const message = error.message
+
+                return send(res, 400, { message })
+            }
+
+            const users = await User.query()
+            const foundUser = await users.find(user => {
+                if (user.id != id && user.email == email) return user
+            })
+
+            if (foundUser) return send(res, 400, { error: 'The email is being used by another user' })
+
+            const passed = await compare(password, user.password)
+            if (!passed) return send(res, 400, { error: 'Incorrect password' })
 
             await User.query().update({
                 name,
@@ -88,26 +130,32 @@ module.exports = {
                 address
             }).where({ id })
 
-            return send(res, 200, { success: 'Atualizado com sucesso!' })
+            return send(res, 200, { success: 'Successfully updated' })
 
         } catch (error) {
-            send(res, 500, { error: 'Erro no servidor.' })
+            send(res, 500, { error: 'Server error' })
         }
     },
 
     async delete(req, res) {
         try {
             const { id } = req.params
-
+            const { password, email } = await json(req)
+            
             let user = await User.query().findById(id)
             if (!user) return send(res, 204, { error: 'Usuário não encontrado' })
 
+            if (email != user.email) return send(res, 400, { error: 'Incorrect email' })
+
+            const passed = await compare(password, user.password)
+            if (!passed) return send(res, 400, { error: 'Incorrect password' })
+
             await User.query().delete().where({ id })
 
-            return send(res, 200, { success: `Usuário ${user.name} deletado com sucesso.` })
+            return send(res, 200, { success: `User ${user.name} successfully deleted.` })
 
         } catch (error) {
-            send(res, 500, { error: 'Erro no servidor.' })
+            send(res, 500, { error: 'Server error' })
         }
     }
 }
